# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DataLens — AI-powered business intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type MetricTrend = typeof MetricTrend[keyof typeof MetricTrend];


export const MetricTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface Metric {
  id: number;
  name: string;
  value: number;
  /** @nullable */
  previousValue?: number | null;
  unit: string;
  category: string;
  trend: MetricTrend;
  date: string;
  createdAt: string;
}

export type MetricInputTrend = typeof MetricInputTrend[keyof typeof MetricInputTrend];


export const MetricInputTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface MetricInput {
  /** @minLength 1 */
  name: string;
  value: number;
  /** @nullable */
  previousValue?: number | null;
  unit: string;
  category: string;
  trend: MetricInputTrend;
  date?: string;
}

export type MetricUpdateTrend = typeof MetricUpdateTrend[keyof typeof MetricUpdateTrend];


export const MetricUpdateTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface MetricUpdate {
  name?: string;
  value?: number;
  /** @nullable */
  previousValue?: number | null;
  unit?: string;
  category?: string;
  trend?: MetricUpdateTrend;
  date?: string;
}

export interface MetricsSummary {
  totalMetrics: number;
  categoriesCount: number;
  avgValue: number;
  upTrendCount: number;
  downTrendCount: number;
  flatTrendCount: number;
  topCategory: string;
}

export interface TrendPoint {
  date: string;
  category: string;
  value: number;
  count: number;
}

export interface CategoryAggregate {
  category: string;
  count: number;
  total: number;
  avg: number;
  upCount: number;
  downCount: number;
}

export interface Dashboard {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface DashboardUpdate {
  name?: string;
  description?: string;
}

export interface DashboardOverview {
  totalDashboards: number;
  totalMetrics: number;
  totalReports: number;
  totalKpiTargets: number;
  kpiHealthPct: number;
  recentMetricsCount: number;
  topPerformingCategory: string;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  performance: 'performance',
  trend: 'trend',
  comparison: 'comparison',
  summary: 'summary',
} as const;

export interface Report {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: ReportType;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  data?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  performance: 'performance',
  trend: 'trend',
  comparison: 'comparison',
  summary: 'summary',
} as const;

export interface ReportInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  type: ReportInputType;
  category?: string;
  data?: string;
}

export type ReportUpdateType = typeof ReportUpdateType[keyof typeof ReportUpdateType];


export const ReportUpdateType = {
  performance: 'performance',
  trend: 'trend',
  comparison: 'comparison',
  summary: 'summary',
} as const;

export interface ReportUpdate {
  title?: string;
  description?: string;
  type?: ReportUpdateType;
  category?: string;
  data?: string;
}

export type KpiTargetStatus = typeof KpiTargetStatus[keyof typeof KpiTargetStatus];


export const KpiTargetStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  behind: 'behind',
  achieved: 'achieved',
} as const;

export interface KpiTarget {
  id: number;
  name: string;
  target: number;
  current: number;
  unit: string;
  category: string;
  status: KpiTargetStatus;
  createdAt: string;
}

export type KpiTargetInputStatus = typeof KpiTargetInputStatus[keyof typeof KpiTargetInputStatus];


export const KpiTargetInputStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  behind: 'behind',
  achieved: 'achieved',
} as const;

export interface KpiTargetInput {
  /** @minLength 1 */
  name: string;
  target: number;
  current: number;
  unit: string;
  category: string;
  status: KpiTargetInputStatus;
}

export type KpiTargetUpdateStatus = typeof KpiTargetUpdateStatus[keyof typeof KpiTargetUpdateStatus];


export const KpiTargetUpdateStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  behind: 'behind',
  achieved: 'achieved',
} as const;

export interface KpiTargetUpdate {
  name?: string;
  target?: number;
  current?: number;
  unit?: string;
  category?: string;
  status?: KpiTargetUpdateStatus;
}

export interface ActivityEvent {
  id: number;
  type: string;
  entityType: string;
  entityId: number;
  description: string;
  createdAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiMessageInput {
  content: string;
}

export type ListMetricsParams = {
category?: string;
limit?: number;
};

export type GetMetricsTrendsParams = {
category?: string;
days?: number;
};

export type ListActivityParams = {
limit?: number;
};