# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DataLens — AI-powered business intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboards:
    get:
      operationId: listDashboards
      tags: [dashboards]
      summary: List all dashboards
      responses:
        "200":
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dashboard"
    post:
      operationId: createDashboard
      tags: [dashboards]
      summary: Create a new dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardInput"
      responses:
        "201":
          description: Created dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
...
  /dashboards/{id}:
    get:
      operationId: getDashboard
      tags: [dashboards]
      summary: Get a single dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
    patch:
      operationId: updateDashboard
      tags: [dashboards]
      summary: Update a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardUpdate"
      responses:
        "200":
          description: Updated dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
    delete:
      operationId: deleteDashboard
      tags: [dashboards]
      summary: Delete a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reports ───────────────────────────────────────────────────────────
...
  /kpi-targets:
    get:
      operationId: listKpiTargets
      tags: [kpi-targets]
      summary: List all KPI targets
      responses:
        "200":
          description: List of KPI targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiTarget"
    post:
      operationId: createKpiTarget
      tags: [kpi-targets]
      summary: Create a KPI target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiTargetInput"
      responses:
        "201":
          description: Created KPI target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiTarget"
  /kpi-targets/{id}:
    patch:
      operationId: updateKpiTarget
      tags: [kpi-targets]
      summary: Update a KPI target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiTargetUpdate"
      responses:
        "200":
          description: Updated KPI target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiTarget"
    delete:
      operationId: deleteKpiTarget
      tags: [kpi-targets]
      summary: Delete a KPI target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Activity Feed ─────────────────────────────────────────────────────
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all AI conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new AI conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message and receive a streaming AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Metric ──────────────────────────────────────────────────────────
...
    Dashboard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    DashboardInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
    DashboardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    KpiTarget:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        target:
          type: number
        current:
          type: number
        unit:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, behind, achieved]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - target
        - current
        - unit
        - category
        - status
        - createdAt
    KpiTargetInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        target:
          type: number
        current:
          type: number
        unit:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, behind, achieved]
      required:
        - name
        - target
        - current
        - unit
        - category
        - status
    KpiTargetUpdate:
      type: object
      properties:
        name:
          type: string
        target:
          type: number
        current:
          type: number
        unit:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, behind, achieved]

    # ── Activity ─────────────────────────────────────────────────────────
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content