# lib/api-zod/src/generated/types/kpiTargetUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DataLens — AI-powered business intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { KpiTargetUpdateStatus } from './kpiTargetUpdateStatus';

export interface KpiTargetUpdate {
  name?: string;
  target?: number;
  current?: number;
  unit?: string;
  category?: string;
  status?: KpiTargetUpdateStatus;
}